chore(experimental): rename concurrency package to load_testing
Disambiguate benchbox/experimental/concurrency (a load-testing +
workload-analysis framework) from the TPC spec concurrency in
core/throughput/, which the old name collided with. Renames the
package to load_testing, renames ConcurrencyAnalyzer to LoadAnalyzer,
repoints coverage config and test imports, and updates README/FAQ
prose. Stays under benchbox.experimental (unsupported, not wired into
CLI/registry) — rename and honest re-documentation only.
