+++ tests/unit/core/load_testing/test_patterns.py
@@ -7,7 +7,7 @@ Licensed under the MIT License. See LICENSE file in the project root for details
 
 import pytest
 
-from benchbox.experimental.concurrency.patterns import (
+from benchbox.experimental.load_testing.patterns import (
     BurstPattern,
     RampUpPattern,
     SpikePattern,
