+++ tests/unit/core/load_testing/test_pool_tester.py
@@ -10,7 +10,7 @@ import time
 
 import pytest
 
-from benchbox.experimental.concurrency.pool_tester import (
+from benchbox.experimental.load_testing.pool_tester import (
     ConnectionAttempt,
     ConnectionPoolTester,
     PoolTestConfig,
